Rename tsrc.github to tsrc.github_client.api_client
